refactor: 불필요한 컴포넌트 삭제 및 통합
