+++ app/(shared)/(ThemeTextField)/createThemeType.ts
@@ -1,6 +1,7 @@
 type ValidationFunction<T> = (value: T) => string;
 
 export type ThemeInfoTextFieldType = {
+  id?: string;
   title?: string;
   content?: string | undefined;
   infoText?: string | undefined;
