+++ app/(shared)/(ThemeTextField)/textField.modules.sass
@@ -1,6 +1,6 @@
-@import '../../../style/variables'
-@import '../../../style/mixins'
-@import '../../../style/button'
+@import '../../style/variables'
+@import '../../style/mixins'
+@import '../../style/button'
 
 .theme-info
   width: 100%
