+++ app/admin-new/(components)/CreateThemeAddButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function CreateThemeAddButton() {
-  return <div className="create-theme__add-button">테마 추가</div>;
+  return <div className="create-theme__add-button disable">테마 추가</div>;
 }
