+++ app/admin-new/(components)/CreateThemeTitle.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { CREATE_THEME_TITLE_TEXT } from "../../(shared)/(ThemeTextField)/(consts)/createTheme";
+import { CREATE_THEME_TITLE_TEXT } from "../../(shared)/(ThemeTextField)/createTheme";
 
 export default function CreateThemeTitle() {
   return <div className="create-theme__title">{CREATE_THEME_TITLE_TEXT}</div>;
