+++ app/admin-new/(style)/createTheme.modules.sass
@@ -34,11 +34,15 @@
     margin-top: 20px
     width: 100%
     height: 58px
-    background-color: $color-white20
+    background-color: $color-white
     box-sizing: border-box
     border-radius: 8px
     text-align: center
     padding: 19px 36px
     @include title18SB
-    color: $color-white20
+    color: $color-black
+
+    &.disable
+      background-color: $color-white20
+      color: $color-white20
